fix: 公演一括作成ページの更新
